Moved the rendering of the dialog up so it animates when you open it
Also removed a dead file!
